Move annotation modal to Annotated dir, remove unused modal
